fix: exclude invoiced payments from invoice total to prevent double-counting
Payments already applied to existing invoices are reflected in the
outstanding invoice remaining_balance. Including them again in
paymentTotal caused the invoice total to be understated (e.g. -$2000
instead of +$1000 for customer 298).

Filter to uninvoiced payments only, matching the existing filter in
dataInsertionOrchestrator.js.

+++ src/endpoints/invoice/invoiceDataInsertions/schemaValidation/invoiceValidation.js
@@ -13,7 +13,7 @@ const invoiceSchema = {
    total_write_offs: 'decimal',
    total_retainers: 'decimal',
    total_amount_due: 'decimal',
-   remaining_balance_on_invoice: 'int',
+   remaining_balance_on_invoice: 'decimal',
    is_invoice_paid_in_full: 'boolean',
    fully_paid_date: ['date', 'null'],
    created_by_user_id: 'int',
